requirements.txt
google-generativeai==0.7.2
pinecone-client==4.1.2
sentence_transformers==3.0.1
numpy==1.26.4
